メッセージの送信失敗をハンドリングする方法を変更

+++ Sources/Konashi/Protocol/NodeElement.swift
@@ -8,5 +8,5 @@
 import Foundation
 
 public protocol NodeElement {
-    var index: Int { get }
+    var index: UInt8 { get }
 }
